moved debug/ to /utils
