+++ manim/__init__.py
@@ -11,7 +11,7 @@ __version__ = importlib_metadata.version(__name__)
 # Importing the config module should be the first thing we do, since other
 # modules depend on the global config dict for initialization.
 from ._config import *
-from .debug import *
+from .utils.debug import *
 
 from .constants import *
